Move I/O code to antsibull.utils.
